Complete rewrite of python wrapper for LibVMI.
PyVMI is built around CFFI, which provides a much cleaner interface between
the C and python code bases.  This includes a libvmi modules that directly
exposes much of the libvmi functionality.  This also includes a pyvmi module
that is more pythonic, allowing for nicer looking python introspection code.
See the example to understand the benefits.

Also updated the volatility address space to work with the new PyVMI.

Finally, this includes a test suite that will run through some basic sanity
checking.  Currently this is hard coded to expect a winxpsp2 VM.  This can
be generalized in the future to allow for broader testing.